requirements.txt
# requirements.txt

# ═══════════════════════════════════════════════════════════════════════════
# CORE FRAMEWORK
# ═══════════════════════════════════════════════════════════════════════════

# Telegram Bot Framework (aiogram 3.x)
aiogram==3.19.0

# Async HTTP client (для Mailgun API)
aiohttp==3.11.16

# Async file operations
aiofiles==24.1.0


# ═══════════════════════════════════════════════════════════════════════════
# DATABASE
# ═══════════════════════════════════════════════════════════════════════════

# SQLAlchemy ORM (2.x)
SQLAlchemy==2.0.40

# PostgreSQL driver (async)
asyncpg==0.30.0

# psycopg2 (sync fallback, binary version)
psycopg2-binary==2.9.10


# ═══════════════════════════════════════════════════════════════════════════
# CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════

# Environment variables
python-dotenv==1.1.0


# ═══════════════════════════════════════════════════════════════════════════
# GOOGLE SERVICES
# ═══════════════════════════════════════════════════════════════════════════

# Google Sheets API
gspread==6.2.0

# Google Auth
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Google API Client
google-api-python-client==2.167.0


# ═══════════════════════════════════════════════════════════════════════════
# EMAIL (Mailgun + SMTP)
# ═══════════════════════════════════════════════════════════════════════════

# Async SMTP client для отправки через свой SMTP сервер
aiosmtplib==3.0.2

# Mailgun работает через HTTP API (используем aiohttp выше)


# ═══════════════════════════════════════════════════════════════════════════
# DOCUMENT GENERATION (PDF, HTML)
# ═══════════════════════════════════════════════════════════════════════════

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF generation from HTML
weasyprint==63.1

# Alternative PDF generation (if needed)
# pdfkit==1.0.0
# wkhtmltopdf binary required separately


# ═══════════════════════════════════════════════════════════════════════════
# BOOKSTACK INTEGRATION
# ═══════════════════════════════════════════════════════════════════════════

# HTTP requests (sync)
requests==2.32.3


# ═══════════════════════════════════════════════════════════════════════════
# CRYPTO & BLOCKCHAIN
# ═══════════════════════════════════════════════════════════════════════════

# Web3 for Ethereum/USDT
web3==7.6.0

# Bitcoin utilities (если проверяем BTC транзакции)
# bitcoinlib==0.6.15

# Algorand (если используем)
# py-algorand-sdk==2.6.0


# ═══════════════════════════════════════════════════════════════════════════
# TASK SCHEDULING
# ═══════════════════════════════════════════════════════════════════════════

# Task scheduler for background jobs
APScheduler==3.10.4


# ═══════════════════════════════════════════════════════════════════════════
# DATE & TIME
# ═══════════════════════════════════════════════════════════════════════════

# Timezone handling
pytz==2024.2

# Date parsing
python-dateutil==2.9.0


# ═══════════════════════════════════════════════════════════════════════════
# UTILITIES
# ═══════════════════════════════════════════════════════════════════════════

# Retry logic for API calls
tenacity==9.0.0


# ═══════════════════════════════════════════════════════════════════════════
# DEVELOPMENT & TESTING (optional, закомментировано)
# ═══════════════════════════════════════════════════════════════════════════

# Testing framework
# pytest==8.3.4
# pytest-asyncio==0.24.0

# Code quality
# black==24.10.0
# flake8==7.1.1
# mypy==1.13.0